appRollback to related component, cleans up the Marathon.jsx
